refactor: Move file copy functions to file_util module
The move improves code organization by grouping related file operations
together in a dedicated module.
